@@ -1,15 +1,16 @@
 name: example
 description: A new Flutter project.
 version: 1.0.0+1
+publish_to: none
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
-  flutter_dotenv: ^2.1.0
+  cupertino_icons: ^1.0.8
+  flutter_dotenv: ^5.2.1
   google_place:
     path: ../
 
