@@ -1,13 +1,13 @@
 name: google_place
-description: A new Flutter package for handle google place api that place search and details and photos and autocomplete and query autocomplete requests are available
-version: 0.4.7
-homepage: https://github.com/bazrafkan/google_place
+description: "A new Flutter package for handling Google Place API: place search, details, photos, autocomplete, and query autocomplete requests."
+version: 0.4.8
+homepage: https://github.com/Rakibul25/google_place
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
-  http: ^0.13.3
+  http: ^1.2.1
 
 dev_dependencies:
   flutter_test:
